+++ src/types/types.ts
@@ -4,7 +4,7 @@ import { HassServiceTarget } from 'home-assistant-js-websocket';
 
 export interface ButtonCardConfig {
   template?: string | string[];
-  triggers_update?: string[] | 'all';
+  triggers_update?: string[] | 'all' | 'update_timer';
   group_expand: boolean;
   type: string;
   entity?: string;
@@ -18,6 +18,9 @@ export interface ButtonCardConfig {
   tap_action?: ActionConfig;
   hold_action?: ActionConfig;
   double_tap_action?: ActionConfig;
+  icon_tap_action?: ActionConfig;
+  icon_hold_action?: ActionConfig;
+  icon_double_tap_action?: ActionConfig;
   show_name?: boolean;
   show_state?: boolean;
   show_icon?: boolean;
@@ -44,6 +47,7 @@ export interface ButtonCardConfig {
   card_size: number;
   tooltip?: string;
   section_mode?: boolean;
+  update_timer?: number;
 }
 
 export interface GridOptions {
@@ -53,7 +57,7 @@ export interface GridOptions {
 
 export interface ExternalButtonCardConfig {
   template?: string | string[];
-  triggers_update?: string[] | 'all';
+  triggers_update?: string[] | 'all' | 'update_timer';
   group_expand?: boolean;
   entity?: string;
   name?: string;
@@ -66,6 +70,9 @@ export interface ExternalButtonCardConfig {
   tap_action?: ActionConfig;
   hold_action?: ActionConfig;
   double_tap_action?: ActionConfig;
+  icon_tap_action?: ActionConfig;
+  icon_hold_action?: ActionConfig;
+  icon_double_tap_action?: ActionConfig;
   show_name?: boolean;
   show_state?: boolean;
   show_icon?: boolean;
@@ -90,6 +97,7 @@ export interface ExternalButtonCardConfig {
   tooltip?: string;
   section_mode?: boolean;
   grid_options?: GridOptions;
+  update_timer?: number;
 }
 
 export type Layout =